chore: add intermediate generator that produce shadcn cli input json
## Background

In out next.js app (the website) we want to:
- show data for each registry item
- show sidebar navigation

For each items we want to:
- add a navigation item in the sidebar
- allow search with a search input box
- show manual install (file content for copy/paste)
- show auto install (cli command for copy/paste)
- show dependencies needed
- show example usage (XXX.example.md)
- show test (XXX.test.ts)

## So what?

Following the same idea of previous commit (please read it), we add a intermediate generation step, that generate json files with everything that the next app needs.

## How it Works

- The source of truth is `registry/input/registry.input.json`, here you add reference to every items of the registry
- You run `pnpm registry:generate-all` that:
    - run `pnpm registry:generate-static-data` to produce the input for shadcn cli, output is `registry/output/static-registry.for-shadcn-cli.json`
    - run `registry:generate-public-files` to create `public/r` file with `shadcn-cli`

The generator is located in `registry/output-generator`

## What is added in this commit

This commit add a new generator in `registry/output-generator/` that produce files in `registry/output`.
No npm script changes, because the script `pnpm registry:generate-all` launchs `registry/output-generator/main.ts`, and this commit add code to this main.ts file.

Changes:
- add `registry/input/config/sidebar-pages-groups-map.json` that contain the map between `registry.input.items[number].categories[0]` and sidebar label
- add code for calculating the json data for next
- add code that print data to file in `registry/output`