chore: fix dynamic import error (module not found)
## Background

dynamyc import in js look at the path of the import only at runtime.
But (i don't know for sure):
- if we are in ts,  and the import path contains a ts alias
- the bundler is webpack or similar to webpack
The path i checked before runtime, producing  a module not found error.

## In our case

In our case we have a conditional import based on APP_DATA_MODE env var.
In case the env var is `static` the path must point to an existing file, if the file doesn't exit, it's correct that Module not Found is triggered.
In case the env var is `compute` the code shouldn't check if the path of `static` conditional points to an existing path. **BUT IT DOES**

## The Fix

The fix come from https://github.com/webpack/webpack/issues/6680#issuecomment-1774110448