chore: now `APP_BASE_URL` constant is computed from other env vars (Vercel preparation)
## Background

This app is never deployed yet.
So to create the build we use a manual APP_BASE_URL env var (in .env file).

Now we are configuring the project so that it can be built on Vercel, and Vercel expose the deploy url with env vars (that are automatically added).
So we remove it from .env file.

## So what

`APP_BASE_URL` is not needed anymore because we get the url from env var exposed by vercel, and in local dev we fallback to http://localhost:3000