refactor: refactor npm script to build public files before build
## Background

On production deploy (Vercel), an item that has as dependency an other item of thi registry, the url of the dependecy contains localhost:3000 as base url.
This, obviously, trigger error.

## Why this happen

Because we generate public files during ddev, then commit them, and during dev the app url is localhost:3000.

## Solution

We move the generation of public files before building next.js.

## What this commit does

- refactored npm script names
- create a new build script that handle everything