fix: now code blocks copy to clipboard copy only the code portion
## Remaining Problem

Ever y item page, in the section `Example`, we render a merkfdown file, that can contain multiple blocks of code.
When this happens (i.e. util-ts-result-zod) the ui is broken because a single box contains all the markdown content.
We need to render code block inside the markdown as separate boxes.