chore: now to render code block we use MarkdownRendererServer direclty
## What you need to know

Beofre this, we use directly these component that internally convert md string to react:
- `CodeCollpasible`
- `CodeNotCollpasible`

After this, we use only `MarkdownRendererServer`,
that internally decide which code component to use based
on md props
```
# CodeCollapsible
\`\`\`ts title=hello.ts isCollpasible=true

# CodeNotCollpasible
\`\`\`ts title=hello.ts
```
## Why

This changes allows as to be more flexible in the `registry-item.example.md` file.
Now we can structure example in indiviudal code blocks like a proper documentation.